update structure for habit util and helper functions
