+++ utils/habits/habitService.ts
@@ -1,0 +1,24 @@
+// service functions for habits
+
+import mmkvStorage from "../mmkvStorage";
+import { HabitObject } from "../types";
+
+export function getHabitObjectFromId(habitId: string) {
+  const activeHabits: HabitObject[] = JSON.parse(
+    mmkvStorage.getString("activeHabits") || "[]"
+  );
+  // looping through the currentActiveHabits array, return the object if found
+  for (const habit of activeHabits) {
+    if (habit.id == habitId) {
+      return habit;
+    }
+  }
+}
+
+export function getAllHabits() {
+  const activeHabits: HabitObject[] = JSON.parse(
+    mmkvStorage.getString("activeHabits") || "[]"
+  );
+
+  return activeHabits;
+}
